corvail-invoices/requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.30.0
google-genai~=1.47.0
pydantic~=2.9.0
pydantic-settings~=2.6.0
httpx~=0.28.1
python-multipart~=0.0.20
sendgrid~=6.11.0
